fix : 탈퇴한 회원의 프로필 조회할 수 있도록 수정 (#224)
